backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx>=0.28.1
requests==2.32.4
pydantic==2.11.7
sqlalchemy==2.0.42
alembic==1.16.4
psycopg2-binary==2.9.9
google-genai==1.28.0
Pillow==11.3.0
google-adk==0.1.0
deprecated==1.2.18 